% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{activeChrs}
\alias{activeChrs}
\title{Match BAM headers bw files and get active chromosome list (from restrict)
(written by Aaron Lun, 12 Dec 2014, copied and modified here)}
\usage{
activeChrs(bam.files, restrict)
}
\arguments{
\item{bam.files}{Character vector (bam files)}

\item{restrict}{Chromosomes to select}
}
\value{
Vector of selected chromosomes
}
\description{
Match BAM headers bw files and get active chromosome list (from restrict)
(written by Aaron Lun, 12 Dec 2014, copied and modified here)
}
