% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_TSS.R
\name{annotate_TSS}
\alias{annotate_TSS}
\title{Annotate detected Transcription Start Sites}
\usage{
annotate_TSS(tssGR, txdb, plot = NA)
}
\arguments{
\item{tssGR}{GRanges object with detected TSS/differential TSS results.}

\item{txdb}{A txdb object.}

\item{plot}{Type of plot to make (choose from "number", "percent" or NA for no plot)}
}
\value{
Annotation of detected TSS
}
\description{
Annotate detected Transcription Start Sites
}
\examples{
# load a txdb object
library("TxDb.Dmelanogaster.UCSC.dm6.ensGene")
seqlevelsStyle(TxDb.Dmelanogaster.UCSC.dm6.ensGene) <- "ENSEMBL"
# annotate a given TSS bed file
tss <- load(file.path(dir, "testTSS_grl.Rdata"))

annotations <- annotate_TSS(tssGR = tss$wt, TxDb.Dmelanogaster.UCSC.dm6.ensGene, plot = "testplot.pdf")

}
