% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_TSS.R
\name{detect_TSS}
\alias{detect_TSS}
\title{Detection of Trancription start sites based on local enrichment}
\usage{
detect_TSS(bam.files, design, foldChange = 2, restrictChr, outfile_prefix)
}
\arguments{
\item{bam.files}{List of bam files to use}

\item{design}{A data frame with rownames = sample names and a column called 'group'
that contains information about the sample group (see example)}

\item{foldChange}{A fold change cutoff of local enrichment to detect the TSS. For samples with
'usual' amount of starting material and squencing depth (>=5ug starting material,
>= 5 mil reads/sample), a cut-off of 6 fold can be used. For samples with low
amount of material or sequencing depth, use a lower cut-off (eg. use 2-fold for
samples with 500ng starting material).}

\item{restrictChr}{Chromosomes to restrict the analysis to.}

\item{outfile_prefix}{Output name prefix for the .bed files}
}
\value{
.bed files containing TSS position for each group, along with a bed file for consensus
	   (union) TSS sites of all samples.
}
\description{
Detection of Trancription start sites based on local enrichment
}
\examples{


}

