% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectDiffTSS.R
\name{diffQCplots}
\alias{diffQCplots}
\title{Make DESeq2 or edgeR QC plots}
\usage{
diffQCplots(method, fit, y, designMat)
}
\arguments{
\item{method}{one of "DESeq2" or "edgeR"}

\item{fit}{output of fitDiffTSS (if method = "edgeR")}

\item{y}{output of fitDiffTSS (if method = "DESeq2")}

\item{designMat}{design matrix (data frame)}
}
\value{
Sparsity, dispersion and PCA plot (if method = DESeq2),
        BCV, dispersion and MDS plot (if method = "edgeR")
}
\description{
Make DESeq2 or edgeR QC plots
}
