% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleCSobject.R
\name{exampleCSobject}
\alias{exampleCSobject}
\title{Create example CapSet object}
\usage{
exampleCSobject(expMethod = "MAPCap")
}
\arguments{
\item{expMethod}{Which experiment method to use (options : "RAMPAGE", "MAPCap")}
}
\value{
An object of class CapSet
}
\description{
Create example CapSet object
}
\examples{
cs <- exampleCSobject(expMethod = "MAPCap")

}
