% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_TSS.R
\name{export_tss}
\alias{export_tss}
\title{Export the detected TSS from CapSet object as .bed files}
\usage{
export_tss(CapSet, pergroup = FALSE, merged = TRUE, outfile_prefix)
}
\arguments{
\item{CapSet}{The modified CapSet object after running \code{\link{detect_TSS}} function}

\item{pergroup}{If TRUE, write output per group of samples}

\item{merged}{If TRUE, write merged bed file (union of all groups)}

\item{outfile_prefix}{Prefix (with path) for output .bed files}
}
\value{
.bed file(s) containing detected TSS.
}
\description{
Export the detected TSS from CapSet object as .bed files
}
\examples{
\dontrun{
export_tss(cs, merged = TRUE, outfile_prefix = "outfolder/wt")
}

}

