% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex_fastq.R
\name{filter_byIDx}
\alias{filter_byIDx}
\title{Filter a read using the borcode in the header}
\usage{
filter_byIDx(idx_name, fq_id, maxM)
}
\arguments{
\item{idx_name}{barcode ID (character)}

\item{fq_id}{fastq read header}

\item{maxM}{max allowd mismatches}
}
\value{
logical vector (which IDs to keep)
}
\description{
Filter a read using the borcode in the header
}

