% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDuplicates.R
\name{filterDups}
\alias{filterDups}
\title{Filter PCR-duplicates from BAM file using internal UMIs}
\usage{
filterDups(bamFile, outFile, keepPairs)
}
\arguments{
\item{bamFile}{character. Input BAM file}

\item{outFile}{character. Output (filtered) BAM file}

\item{keepPairs}{logical. Keep R2 read?}
}
\value{
Filtered BAM file, after PCR duplicate removal
}
\description{
Filter PCR-duplicates from BAM file using internal UMIs
}
