% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_diffTSS.R
\name{fit_diffTSS}
\alias{fit_diffTSS}
\title{Detect differentially expressed Transcription Start Sites between two conditions (fit model)}
\usage{
fit_diffTSS(CSobject, TSSfile, groups, outplots = NULL, plotref)
}
\arguments{
\item{CSobject}{An object of class \code{\link{CapSet}}}

\item{TSSfile}{A .bed file with TSS positions to test for differential TSS analysis. If left empty,
the union of detected TSS present within the provided CSobject would be plotted.}

\item{outplots}{Output pdf filename for plots. If provided, the plots for BCV, dispersion and
MDS plot is created and saved in this file.}

\item{plotref}{Name of reference sample to plot for detection of composition bias in the
data. Data is normalized using the TMM method to avoid composition bias.}

\item{design}{A data frame with rownames = sample names and a column called 'group'
that contains information about the sample group (see example)}
}
\value{
Returns an object of class DGEGLM.
}
\description{
Detect differentially expressed Transcription Start Sites between two conditions (fit model)
}
\examples{


}

