% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneCounts.R
\name{get_geneCounts}
\alias{get_geneCounts}
\title{Get gene-level counts from TSS data}
\usage{
get_geneCounts(transcriptGRL, bamfiles, regionAroundTSS = 500,
  single_end = TRUE, outfile = NA)
}
\arguments{
\item{transcriptGRL}{A GRangesList object containing transcripts, created using transcriptsBy(txdb)}

\item{bamfiles}{Bam files to count the reads from}

\item{regionAroundTSS}{How many bases downstream of TSS to count}

\item{single_end}{Logical, indicating whether reads are single end}

\item{outfile}{Tab-separated output file name (if required)}
}
\value{
data.frame with gene-level counts for all genes in the txdb object
}
\description{
Get gene-level counts from TSS data
}
\examples{

 # load a txdb object
 dm6gtf <- loadDB(system.file("extdata/dm6_GTF.DB", package = "icetea"))
 # get transcripts by gene
 dm6trans <- transcriptsBy(dm6gtf, "gene")
 # get gene counts, counting reads around 500 bp of the TSS
 gcounts <- get_geneCounts(dm6trans, bamfiles)

}

