% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplexFASTQ.R
\name{get_newfastq}
\alias{get_newfastq}
\title{Get data to create new ShortReadQ object after barcode trimming}
\usage{
get_newfastq(type, fq_R1, fq_R2)
}
\arguments{
\item{type}{character. expType of the CapSet object}

\item{fq_R1}{character. fastq Read 1}

\item{fq_R2}{character. fastq Read 2}
}
\value{
A list with new R1 and R2 sequence, quality, barcode string and sample id
}
\description{
Get data to create new ShortReadQ object after barcode trimming
}
