% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBamFlags}
\alias{getBamFlags}
\title{Get flags to read from bam}
\usage{
getBamFlags(countAll)
}
\arguments{
\item{paired}{logical. Keep Paired reads?}
}
\value{
bamFlags
}
\description{
Get flags to read from bam
}
