% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChromBins}
\alias{getChromBins}
\title{Get chromosome bins from BAM files}
\usage{
getChromBins(bamFiles, restrictChr = NULL, binSize)
}
\arguments{
\item{bamFiles}{Character. bam files}

\item{restrictChr}{character. Chromosomes to select}

\item{binSize}{numeric. Size of bins}
}
\value{
GRanges (bins) for both strands
}
\description{
Get chromosome bins from BAM files
}
