% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChromWindows}
\alias{getChromWindows}
\title{Get chromosome sliding windows from BAM files}
\usage{
getChromWindows(bamFiles, restrictChr = NULL, binSize, stepSize)
}
\arguments{
\item{bamFiles}{Character vector (bam files)}

\item{restrictChr}{Chromosomes to select}

\item{binSize}{Size of bins}

\item{stepSize}{Size of window slide}
}
\value{
GRanges (sliding windows) for both strands
}
\description{
Get chromosome sliding windows from BAM files
}
