% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMCparams}
\alias{getMCparams}
\title{Get platform-specific multicore params}
\usage{
getMCparams(cores)
}
\arguments{
\item{cores}{No. of cores}
}
\value{
BPPARAM
}
\description{
Get platform-specific multicore params
}
