% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_TSS.R
\name{getranks}
\alias{getranks}
\title{Assign feature ranks on a VariantAnnotation output}
\usage{
getranks(x, rank_vec)
}
\arguments{
\item{x}{output from VariantAnnotation}

\item{rank_vec}{the pre-set vector of ranks}
}
\value{
A vector of ranks of length = length of input features
}
\description{
Assign feature ranks on a VariantAnnotation output
}
