% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{localFilter}
\alias{localFilter}
\title{Calculate local enrichment of windows over background
'local' filter copied and modified from csaw::filterWindows
written by Aaron Lun (5 November 2014, last modified 3 March 2017)}
\usage{
localFilter(data, background, assay.data = 1, assay.back = 1)
}
\arguments{
\item{data}{RangedSE object (windows)}

\item{background}{RangedSE object (background)}

\item{assay.data}{Arg to pass forward}

\item{assay.back}{Arg to pass forward}
}
\value{
list with data and background abundances
}
\description{
Calculate local enrichment of windows over background
'local' filter copied and modified from csaw::filterWindows
written by Aaron Lun (5 November 2014, last modified 3 March 2017)
}
