% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCaps.R
\name{mapCaps}
\alias{mapCaps}
\title{Map the data from 5' profiling techniques}
\usage{
mapCaps(CSobject, genomeIndex, outdir, externalGTF = NULL, nthreads = 1,
  logfile = NULL)
}
\arguments{
\item{CSobject}{An object of class \code{\link{CapSet}}}

\item{genomeIndex}{path to the Subread index file. Should end with the basename of the index.}

\item{outdir}{output directory path}

\item{externalGTF}{(optional) provide external annotation file in `GTF`` format to
increase alignment accuracy}

\item{nthreads}{number of threads to use for mapping.}

\item{logfile}{a log file to write the processing message.}
}
\value{
modified CapSet object with mapping information. Mapped and sorted BAM
        files are saved in `outdir`.
}
\description{
Map the data from 5' profiling techniques
}
\examples{
\dontrun{
# before mapping :
# 1. Create a CapSet object
# 2. de-multiplex the fastqs

# load a previously saved CapSet object
cs <- exampleCSobject()

# map the data (not available on windows)
library(Rsubread)
dir.create("bam")
buildindex(basename = "dm6", reference = "/path/to/dm6_genome.fa")
cs <- mapCaps(cs, genomeIndex = "dm6", outdir = "bam", nthreads = 10)

}

}
