% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\docType{methods}
\name{plot_TSSprecision}
\alias{plot_TSSprecision}
\alias{plot_TSSprecision,CapSet-method}
\title{Compare the precision of TSS detection between multiple samples}
\usage{
plot_TSSprecision(CSobject, ...)

plot_TSSprecision(CSobject, ...)

\S4method{plot_TSSprecision}{CapSet}(CSobject, reference,
  distanceCutoff = 500, outFile = NULL)
}
\arguments{
\item{reference}{Reference Transcrips/Genes as a \code{\link{GRanges}} object}

\item{distanceCutoff}{Maximum distance (in base pairs) from reference TSS to plot}

\item{outFile}{Output file name (filename extention would be used to determine type).
If outfile not specified, the plot would be retured on the screen}

\item{TSSbedFiles}{BED files with TSS called by the CAGE wrapper}

\item{sampleNames}{Labels for Input samples}

\item{object}{An object of class \code{\link{CapSet}}}

\item{reference}{Reference Transcrips/Genes as a \code{\link{GRanges}} object}

\item{distanceCutoff}{Maximum distance (in base pairs) from reference TSS to plot}

\item{outFile}{Output file name (filename extention would be used to determine type).
If outfile not specified, the plot would be retured on the screen}
}
\value{
A ggplot object, or a file. Plot showing perent of TSS detected per sample with respect to
        their cumulative distance to TSS of the provided reference
}
\description{
Plot precision of TSS detection from multiple samples (bed files) with respect to
a given reference annotation.

Plot precision of TSS detection from multiple samples present within a \
code{\link{CapSet}} object, with respect to a given reference annotation.
}
\examples{
\dontrun{
library("TxDb.Dmelanogaster.UCSC.dm6.ensGene")
transcripts <- transcripts(dm6GTF)

Plotting the precision using a pre computed set of TSS (.bed files) :

files <- system.file("extdata", "testTSS.bed", package = "mapcapR")
plot_TSSprecision(TSSbedFiles = files, sampleNames = "testTSS",
		reference = transcripts , distanceCutoff = 500,
		outFile = "TSS_detection_precision.png")
}
\dontrun{
Plotting the precision using a CapSet object :

library("TxDb.Dmelanogaster.UCSC.dm6.ensGene")
transcripts <- transcripts(dm6GTF)

plot_TSSprecision(CapSet = cs, reference = transcripts, outFile = "TSS_detection_precision.png")
}

}

