% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotPrecision}
\alias{plotPrecision}
\title{Plotprecision background script}
\usage{
plotPrecision(ref, tssData, distCut)
}
\arguments{
\item{ref}{reference GRanges}

\item{tssData}{GRangesList object with TSS detected per sample}

\item{distCut}{max distance cutoff}
}
\value{
ggplot object
}
\description{
Plotprecision background script
}

