% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readsToCenter}
\alias{readsToCenter}
\title{preprocess reads to count only center overlaps}
\usage{
readsToCenter(reads, width = 1, fix = "center")
}
\arguments{
\item{reads}{GAlignment object to resize}

\item{width}{integer. New read length}

\item{fix}{character. 'Start' for 5'}
}
\value{
Resized reads as GRanges
}
\description{
preprocess reads to count only center overlaps
}
