% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ResizeReads}
\alias{ResizeReads}
\title{preprocess reads to count only 5' overlaps}
\usage{
ResizeReads(reads, width = 1, fix = "start")
}
\arguments{
\item{reads}{GAlignment object}

\item{width}{New read length}

\item{fix}{'Start' for 5'}
}
\value{
Resized reads as GRanges
}
\description{
preprocess reads to count only 5' overlaps
}
