% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{sampleInfo,CapSet-method}
\alias{sampleInfo,CapSet-method}
\title{get sample information data frame}
\usage{
\S4method{sampleInfo}{CapSet}(object)
}
\arguments{
\item{object}{the \code{\link{CapSet}} object}
}
\description{
get sample information data frame
}

