% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{sampleInfo<-,CapSet-method}
\alias{sampleInfo<-,CapSet-method}
\title{reset sample information data frame}
\usage{
\S4method{sampleInfo}{CapSet}(object) <- value
}
\arguments{
\item{object}{The \code{\link{CapSet}} object}

\item{value}{The replacement sampleInfo data frame}
}
\value{
None
}
\description{
reset sample information data frame
}
\examples{
# load a previously saved CapSet object
cs <- exampleCSobject()
# get sampleinfo
si <- sampleInfo(cs)
# modify
si$samples <- paste0("sample_", 1:6)
# replace
sampleInfo(cs) <- si


}

