% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sampleInfo<-}
\alias{sampleInfo<-}
\title{reset sample information data frame}
\usage{
sampleInfo(object, ...) <- value
}
\arguments{
\item{object}{The \code{\link{CapSet}} object}

\item{...}{...}

\item{value}{new value}
}
\description{
reset sample information data frame
}

