% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sampleInfo}
\alias{sampleInfo}
\title{get sample information data frame}
\usage{
sampleInfo(object, ...)
}
\arguments{
\item{object}{the \code{\link{CapSet}} object}

\item{...}{...}
}
\value{
sample information data.frame
}
\description{
get sample information data frame
}
\examples{

# load a previously saved CapSet object
cs <- exampleCSobject()
# get sampleinfo
si <- sampleInfo(cs)

}
