% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplexFASTQ.R
\name{split_fastq}
\alias{split_fastq}
\title{Split paired-end fastq by barcodes}
\usage{
split_fastq(expType, idx_name, outfile_R1, outfile_R2, fastq_R1, fastq_R2,
  max_mismatch)
}
\arguments{
\item{expType}{character. experiment type (RAMPAGE, MAPCap or CAGE)}

\item{idx_name}{character. barcode ID}

\item{outfile_R1}{character. output fastq file : Read 1}

\item{outfile_R2}{character. output fastq file : Read 2}

\item{fastq_R1}{character. input fastq file : Read 1}

\item{fastq_R2}{character. input fastq file : Read 2}

\item{max_mismatch}{integer. max allowed mismatches}
}
\value{
kept reads corresponding to each barcode.
}
\description{
Split paired-end fastq by barcodes
}
