% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateTSS.R
\name{splitranks}
\alias{splitranks}
\title{Get features with the best rank for each TSS}
\usage{
splitranks(x)
}
\arguments{
\item{x}{output of getranks}
}
\value{
A data frame with counts
}
\description{
Get features with the best rank for each TSS
}
