% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_TSS.R
\name{strandBinCounts}
\alias{strandBinCounts}
\title{Perform stranded Bin counts}
\usage{
strandBinCounts(bam.files, restrictChrs, bam_param, bp_param, window_size,
  sliding = FALSE)
}
\arguments{
\item{bam.files}{character vector. BAM files to use}

\item{restrictChrs}{character vector. chromosomes to use}

\item{bam_param}{ScanBAMParams}

\item{bp_param}{BPPARAM}

\item{window_size}{integer. size of window to use}

\item{sliding}{logical. perform sliding window counts?}
}
\value{
RangedSE object with forward and reverse strand counts
}
\description{
Perform stranded Bin counts
}
