% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimFastqIndex.R
\name{trimFastqIndex}
\alias{trimFastqIndex}
\title{Trim the sample indexes off the fastq files (RAMPAGE/MAPCap)}
\usage{
trimFastqIndex(CapSet, outdir)
}
\arguments{
\item{CapSet}{The CapSet object}

\item{outdir}{output directory to keep the trimmed fastq files}
}
\value{
fastq files
}
\description{
Trim the sample indexes off the fastq files (RAMPAGE/MAPCap)
}
\examples{

}

