% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_TSS.R
\name{annotate_TSS}
\alias{annotate_TSS}
\title{Annotate detected Transcription Start Sites}
\usage{
annotate_TSS(tssFile, txdb, plot = NA)
}
\arguments{
\item{tssFile}{BED file with detected TSS/differential TSS results.}

\item{txdb}{A txdb object.}

\item{plot}{Type of plot to make (choose from "number", "percent" or NA for no plot)}
}
\value{
Annotation of detected TSS
}
\description{
Annotate detected Transcription Start Sites
}
\examples{


}

