% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex_fastq.R
\name{demultiplex_fastq}
\alias{demultiplex_fastq}
\title{Demultiplex (tagged) fastq files using sample barcodes}
\usage{
demultiplex_fastq(CSobject, max_mismatch, outdir, ncores = 1)
}
\arguments{
\item{CSobject}{CapSet object created using \code{\link{newCapSet}} function}

\item{max_mismatch}{maximum allowd mismatches}

\item{outdir}{path to output directory}

\item{ncores}{No. of cores to use}
}
\value{
de-multiplxed fastq files corresponding to each barcode. The files are written
        on disk with the corresponding sample names as specified in sampleBarcodes(CapSet)
}
\description{
Demultiplex (tagged) fastq files using sample barcodes
}
\examples{
# load a previously saved CapSet object
cs <- exampleCSobject()
# demultiplex allowing one mismatch in sample indexes
cs <- demultiplex_fastq(cs, outdir = dir, max_mismatch = 1)

}
