% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex_fastq.R
\name{demultiplex_fastq}
\alias{demultiplex_fastq}
\title{Demultiplex (tagged) fastq files using sample barcodes}
\usage{
demultiplex_fastq(CapSet, max_mismatch, outdir, nthreads = 1)
}
\arguments{
\item{CapSet}{CapSet object created using \code{\link{newCapSet}} function}

\item{max_mismatch}{maximum allowd mismatches}

\item{outdir}{path to output directory}

\item{nthreads}{No. of threads to use}
}
\value{
de-multiplxed fastq files corresponding to each barcode. The files are written
        on disk with the corresponding sample names as specified in sampleBarcodes(CapSet)
}
\description{
Demultiplex (tagged) fastq files using sample barcodes
}
\examples{
\dontrun{
idxlist <- c("CAAGTG", "TTAGCC", "GTGGAA", "GGTTAC", "TGTGAG", "CATCAC")
fnames <- c("GFPa","GFPb","BEAF32a","BEAF32b","BEAF32Rrp6a","BEAF32Rrp6b")

cs <- newCapSet(expMethod = 'MAPCap', fastqType = 'paired',
	fastq_R1 = '../../2017_MAPCap/MAPCap_F/01_fastq/trimmed/MAPCap_F_R1.fastq.gz',
	fastq_R2 = '../../2017_MAPCap/MAPCap_F/01_fastq/trimmed/MAPCap_F_R2.fastq.gz',
	sampleInfo = data.frame(row.names = idxlist, files = fnames))

system.time(demultiplex_fastq(cs, max_mismatch = 2))
}
}

