% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_diffTSS.R
\name{detect_diffTSS}
\alias{detect_diffTSS}
\title{Detect differentially expressed Transcription Start Sites between two conditions (test)}
\usage{
detect_diffTSS(fit, testGroup, contGroup)
}
\arguments{
\item{fit}{DGEGLM object (output of \code{\link{fit_diffTSS}} command )}

\item{testGroup}{Test group name}

\item{contGroup}{Control group name}
}
\value{
A \code{\link{GRanges}} object containing p-values of differential expression for each TSS.
}
\description{
Detect differentially expressed Transcription Start Sites between two conditions (test)
}
\examples{


}

