% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_diffTSS.R
\name{detect_diffTSS}
\alias{detect_diffTSS}
\title{Detect differentially expressed Transcription Start Sites between two conditions (test)}
\usage{
detect_diffTSS(fit, testGroup, contGroup, TSSfile, MAplot_fdr = NA)
}
\arguments{
\item{fit}{DGEGLM object (output of \code{\link{fit_diffTSS}} command )}

\item{testGroup}{Test group name}

\item{contGroup}{Control group name}

\item{TSSfile}{The TSS .bed file used for \code{\link{fit_diffTSS}} command}

\item{MAplot_fdr}{FDR threshold to mark differentially expressed TSS in MAplot (NA = Don't make an MAplot)}
}
\value{
A \code{\link{GRanges}} object containing p-values of differential expression for each TSS.
}
\description{
Detect differentially expressed Transcription Start Sites between two conditions (test)
}
\examples{

# before running this
# 1. Create a CapSet object
# 2. de-multiplex the fastqs
# 3. map them
# 4. filter duplicate reads from mapped BAM
# 5. detect TSS
# 6. fit the diff TSS model.

\dontrun{
# load a previously saved DGEGLM object from step 5
csfit <- load("diffTSS_fit.Rdata")

# detect differentially expressed TSS between groups (return MA plot)
detect_diffTSS(csfit, testGroup = "mut", controlGroup = "wt",
               tssFile = "testTSS_merged.bed", MAplot_fdr = 0.05)

}

}

