% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_diffTSS.R
\name{detect_diffTSS}
\alias{detect_diffTSS}
\title{Detect differentially expressed Transcription Start Sites between two conditions (test)}
\usage{
detect_diffTSS(fit, testGroup, contGroup, TSSfile, MAplot_fdr = NA)
}
\arguments{
\item{fit}{DGEGLM object (output of \code{\link{fit_diffTSS}} command )}

\item{testGroup}{Test group name}

\item{contGroup}{Control group name}

\item{MAplot_fdr}{FDR threshold to mark differentially expressed TSS in MAplot (NA = Don't make an MAplot)}

\item{tssFile}{The TSS .bed file used for \code{\link{fit_diffTSS}} command}
}
\value{
A \code{\link{GRanges}} object containing p-values of differential expression for each TSS.
}
\description{
Detect differentially expressed Transcription Start Sites between two conditions (test)
}
\examples{


}

