% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_TSS.R
\name{detect_TSS}
\alias{detect_TSS}
\title{Detection of Trancription start sites based on local enrichment}
\usage{
detect_TSS(CapSet, groups, outfile_prefix = NULL, foldChange = 2,
  restrictChr = NULL)
}
\arguments{
\item{CapSet}{CapSet object created using \code{\link{newCapSet}} function}

\item{groups}{a character vector that contains group name of the sample, for replicate-based TSS
calling (see example)}

\item{outfile_prefix}{Output name prefix for the .Rdata file containing window counts, background counts
and filtering statistics calculated during TSS detection.}

\item{foldChange}{A fold change cutoff of local enrichment to detect the TSS. For samples with
'usual' amount of starting material and squencing depth (>=5ug starting material,
>= 5 mil reads/sample), a cut-off of 6 fold can be used. For samples with low
amount of material or sequencing depth, use a lower cut-off (eg. use 2-fold for
samples with 500ng starting material).}

\item{restrictChr}{Chromosomes to restrict the analysis to.}
}
\value{
.bed files containing TSS position for each group, along with a bed file for consensus
	   (union) TSS sites of all samples.
}
\description{
Detection of Trancription start sites based on local enrichment
}
\examples{
\dontrun{
cs <- detect_TSS(Capset = cs, group = c("wt","wt","mut","mut"), outfile_prefix = "testTSS",
           foldChange = 6, restrictChr = c("2L","2R","X"))
}

}

