% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSobject_example.R
\name{exampleCSobject}
\alias{exampleCSobject}
\title{Create example CapSet object}
\usage{
exampleCSobject(protocol = "MAPCap")
}
\value{
An object of class CapSet
}
\description{
Create example CapSet object
}
\examples{
cs <- exampleCSobject(protocol = "MAPCap")

}

