% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDuplicates.R
\name{filterDuplicates}
\alias{filterDuplicates}
\title{Filter PCR-duplicates from mapped files using internal UMIs}
\usage{
filterDuplicates(CapSet, outdir)
}
\arguments{
\item{CapSet}{an object of class \code{\link{CapSet}}}

\item{outdir}{output directory for filtered BAM files}
}
\value{
modified CapSet object with filtering information. Filtered BAM files are saved in `outdir`.
}
\description{
This script considers the read mapping start position and the UMI to determine whether a
             read is a PCR duplicate. All PCR duplicates are then removed and one entry per read is kept.
             In case of paired-end reads (MAPCap/RAMPAGE), only one end (R1) is kept after filtering.
}

