% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/filterDuplicates.R
\docType{methods}
\name{filterDuplicates}
\alias{filterDuplicates}
\alias{filterDuplicates,CapSet-method}
\title{Filter PCR-duplicates from mapped files using internal UMIs}
\usage{
filterDuplicates(CSobject, outdir)

\S4method{filterDuplicates}{CapSet}(CSobject, outdir)
}
\arguments{
\item{CSobject}{an object of class \code{\link{CapSet}}}

\item{outdir}{output directory for filtered BAM files}
}
\value{
modified CapSet object with filtering information. Filtered BAM files are saved in `outdir`.
}
\description{
This script considers the read mapping start position and the UMI to determine whether a
             read is a PCR duplicate. All PCR duplicates are then removed and one entry per read is kept.
             In case of paired-end reads (MAPCap/RAMPAGE), only one end (R1) is kept after filtering.
}
\examples{

# before running this
# 1. Create a CapSet object
# 2. de-multiplex the fastqs
# 3. map them

# load a previously saved CapSet object
cs <- exampleCSobject()
# filter duplicate reads from mapped BAM files
dir.create("filtered_bam")
cs <- filterDuplicates(cs, outdir = "filtered_bam")

}
