% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDups.R
\name{filterDuplicates}
\alias{filterDuplicates}
\title{Filter PCR-duplicates from BAM file using internal UMIs}
\usage{
filterDuplicates(bamFile, outFile)
}
\arguments{
\item{bamFile}{Input BAM file}

\item{outFile}{Output (filtered) BAM file}
}
\value{
Filtered BAM file (with only R1), after PCR duplicate removal
}
\description{
This script considers the read mapping start posion and the UMI to determine whether a
             read is a PCR duplicate. All PCR duplicates are then removed and one entry per read is kept.
             In case of paired-end reads (MAPCap/RAMPAGE), only one end (R1) is kept after filtering.
}
\examples{

bam <- system.file("extdata", "test_mapped.bam", package = "mapcapR")
filterDuplicates(bamFile = bam, outFile = "test_rmDup.bam")

}

