% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDuplicates.R
\name{filterDups}
\alias{filterDups}
\title{Filter PCR-duplicates from BAM file using internal UMIs}
\usage{
filterDups(bamFile, outFile)
}
\arguments{
\item{bamFile}{Input BAM file}

\item{outFile}{Output (filtered) BAM file}
}
\value{
Filtered BAM file (with only R1), after PCR duplicate removal
}
\description{
Filter PCR-duplicates from BAM file using internal UMIs
}

