% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_diffTSS.R
\name{fit_diffTSS}
\alias{fit_diffTSS}
\title{Detect differentially expressed Transcription Start Sites between two conditions (fit model)}
\usage{
fit_diffTSS(bam.files, TSSfile, design, outplots, plotref)
}
\arguments{
\item{bam.files}{Character vector with paths of bam files to use}

\item{TSSfile}{A .bed file with TS sites to test. Normally it would be the *merged.bed file
(output of \code{\link{detect_TSS}} command)}

\item{design}{A data frame with rownames = sample names and a column called 'group'
that contains information about the sample group (see example)}

\item{outplots}{Output pdf filename for plots}

\item{plotref}{Name of reference sample to plot for detection of composition bias in the
data. Data is normalized using the TMM method to avoid composition bias.}
}
\value{
Returns an object of class DGEGLM.
}
\description{
Detect differentially expressed Transcription Start Sites between two conditions (fit model)
}
\examples{


}

