% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneCounts.R
\name{get_geneCounts}
\alias{get_geneCounts}
\title{Get gene-level counts from TSS data}
\usage{
get_geneCounts(txdb, bamfiles, outfile = NA)
}
\arguments{
\item{txdb}{A txdb object}

\item{bamfiles}{Bam files to count the reads from}

\item{outfile}{Tab-separated output file name (if required)}
}
\value{
data.frame with gene-level counts for all genes in the txdb object
}
\description{
Get gene-level counts from TSS data
}
\examples{

}

