% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneCounts.R
\name{get_geneCounts}
\alias{get_geneCounts}
\title{Get gene-level counts from TSS data}
\usage{
get_geneCounts(CSobject, transcriptGRL, regionAroundTSS = 500,
  single_end = TRUE, outfile = NA)
}
\arguments{
\item{CSobject}{The \code{\link{CapSet}} object to use.}

\item{transcriptGRL}{A GRangesList object containing transcripts, created using transcriptsBy(txdb)}

\item{regionAroundTSS}{How many bases downstream of TSS to count}

\item{single_end}{Logical, indicating whether reads are single end}

\item{outfile}{Tab-separated output file name (if required)}
}
\value{
data.frame with gene-level counts for all genes in the txdb object
}
\description{
Get gene-level counts from TSS data
}
\examples{

 # load a txdb object
 library("TxDb.Dmelanogaster.UCSC.dm6.ensGene")
 seqlevelsStyle(TxDb.Dmelanogaster.UCSC.dm6.ensGene) <- "ENSEMBL"
 # get transcripts by gene
 dm6trans <- transcriptsBy(TxDb.Dmelanogaster.UCSC.dm6.ensGene, "gene")
 # load a CapSet object
 cs <- exampleCSobject()
 # get gene counts, counting reads around 500 bp of the TSS
 gcounts <- get_geneCounts(cs, dm6trans)

}
