% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex_fastq.R
\name{get_newfastq}
\alias{get_newfastq}
\title{Filter a read using the borcode in the header}
\usage{
get_newfastq(type, fq_R1, fq_R2)
}
\arguments{
\item{type}{expType of the CapSet object}

\item{fq_R1}{fastq Read 1}

\item{fq_R2}{fastq Read 2}

\item{idx_name}{barcode ID (character)}

\item{fq_id}{fastq read header}

\item{maxM}{max allowd mismatches}
}
\value{
logical vector (which IDs to keep)

Get data to create new ShortReadQ object after barcode trimming

A list with new R1 and R2 sequence, quality, barcode string and sample id
}
\description{
Filter a read using the borcode in the header
}

