% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_TSS.R
\name{getranks}
\alias{getranks}
\title{Assign feature ranks on a VariantAnnotation output}
\usage{
getranks(x)
}
\arguments{
\item{x}{output from VariantAnnotation}
}
\value{
A list of ranks
}
\description{
Assign feature ranks on a VariantAnnotation output
}

