% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCaps.R
\name{mapCaps}
\alias{mapCaps}
\title{Map the CAGE data}
\usage{
mapCaps(CapSet, genomeIndex, outdir, nthreads, logfile = NULL, ...)
}
\arguments{
\item{CapSet}{An object of class \code{\link{CapSet}}}

\item{genomeIndex}{path to the Subread index file. Should end with the basename of the index.}

\item{outdir}{output directory path}

\item{nthreads}{number of threads to use for mapping.}

\item{logfile}{a log file to write the processing message.}

\item{...}{additional arguments passed to the \code{\link{subjunc}} function.}
}
\value{
modified CapSet object with mapping information. Mapped and sorted bam files are saved in `outdir`.
}
\description{
Map the CAGE data
}
\examples{


}

