% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCaps.R
\name{mapCaps}
\alias{mapCaps}
\title{Map the paired-end MAPCap data.}
\usage{
mapCaps(index, R1, R2, outprefix, nthreads, logfile = NULL, ...)
}
\arguments{
\item{index}{character string giving the basename of the Subread index file.}

\item{R1}{forward read (_R1) fastq, output of \code{\link{trimFastqIndex}} command.}

\item{R2}{reverse read (_R2) fastq, output of \code{\link{trimFastqIndex}} command.}

\item{outprefix}{output file prefix (without ".bam" extention)}

\item{nthreads}{number of threads to use for mapping.}

\item{logfile}{a log file to write the processing message.}

\item{...}{additional arguments passed to the \code{\link{subjunc}} function.}
}
\value{
bamfile A mapped BAM file for the sample.
}
\description{
Map the paired-end MAPCap data.
}
\examples{
r1 <- system.file("extdata", "testout_R1.fastq.gz", package = "mapcapR")
r2 <- system.file("extdata", "testout_R2.fastq.gz", package = "mapcapR")
\dontrun{
mapCaps(index,R1 = r1, R2 = r2, output = "test_mapped", nthreads = 10, logfile=NULL)
}

}

