% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{newCapSet}
\alias{newCapSet}
\title{Create a new CapSet object}
\usage{
newCapSet(expMethod, fastqType, fastq_R1, fastq_R2 = NA, sampleBarcodes)
}
\arguments{
\item{expMethod}{experiment method ('CAGE', 'RAMPAGE' or 'MAPCap')}

\item{fastqType}{fastq file type. Should be either 'single' or 'paired'}

\item{fastq_R1}{path for Read R1 (or file path for single end reads)}

\item{fastq_R2}{path for Read R2 (for paired end reads)}

\item{sampleBarcodes}{for de-multiplexing, a data.frame with two columns. First column should have the barcodes,
while the second column should have the corresponding sample names.}
}
\value{
An object of class CapSet
}
\description{
Create a new CapSet object
}
\examples{

\dontrun{
newCapSet(fastqType = "single", fastq_R1 = "out_R1.fastq.gz",
fastq_R2 = "out_R2.fastq.gz" , expMethod = "RAMPAGE", sampleBarcodes = df)

}

}

