% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{newCapSet}
\alias{CapSet}
\alias{CapSet-class}
\alias{newCapSet}
\title{Create a new CapSet object}
\usage{
newCapSet(expMethod, fastqType, fastq_R1, fastq_R2 = NULL, sampleInfo)
}
\arguments{
\item{expMethod}{experiment method ('CAGE', 'RAMPAGE' or 'MAPCap')}

\item{fastqType}{fastq file type. Should be either 'single' or 'paired'}

\item{fastq_R1}{path for Read R1 (or file path for single end reads)}

\item{fastq_R2}{path for Read R2 (for paired end reads)}

\item{sampleInfo}{data.frame with sample information, with row.names = barcode sequences,
and first column containing corresponding sample names.}
}
\value{
An object of class CapSet
}
\description{
Create a new CapSet object

CapSet object
}
\examples{

\dontrun{
df <- data.frame(row.names = c("TTAGCC" ,"CAAGTG"), samples = c("one", "two") )
cs <- newCapSet(fastqType = "paired",
                fastq_R1 = "inst/extdata/test_R1.fastq.gz",
                fastq_R2 = "inst/extdata/test_R2.fastq.gz",
                expMethod = "MAPCap", sampleInfo = df)
}

}

