% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{newCapSet}
\alias{CapSet}
\alias{CapSet-class}
\alias{newCapSet}
\title{Create a new CapSet object}
\usage{
newCapSet(sampleInfo, expMethod, fastqType, fastq_R1, fastq_R2 = NULL)
}
\arguments{
\item{sampleInfo}{data.frame with sample information, with row.names = barcode sequences,
and first column containing corresponding sample names.}

\item{expMethod}{experiment method ('CAGE', 'RAMPAGE' or 'MAPCap')}

\item{fastqType}{fastq file type. Should be either 'single' or 'paired'}

\item{fastq_R1}{path for Read R1 (or file path for single end reads)}

\item{fastq_R2}{path for Read R2 (for paired end reads)}
}
\value{
An object of class CapSet
}
\description{
Create a new CapSet object

CapSet object
}
\examples{

# list of barcode IDs
idxlist <- c("CAAGTG", "AGATGC", "TGTGAG",
		 "GGTTAC", "TTAGCC", "AGTCGA")
# corresponding sample names
fnames <-  c("WTa", "WTb", "WTc",
		 "MLEa", "MLEb", "MLEc")

# create a new capset object and save
dir <- system.file("extdata", package="icetea")
cs <- newCapSet(expMethod = 'MAPCap', fastqType = 'paired',
	fastq_R1 = file.path(dir, 'mapcap_test_R1.fastq.gz'),
	fastq_R2 = file.path(dir, 'mapcap_test_R2.fastq.gz'),
	sampleInfo = data.frame(row.names = idxlist, files = fnames))

save(cs, file = file.path(dir, "CSobject.Rdata") )

}

