% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numReadsInBed}
\alias{numReadsInBed}
\title{Count the number of reads in a given GRanges}
\usage{
numReadsInBed(regions, bams = NA, countall = FALSE)
}
\arguments{
\item{regions}{The GRanges object}

\item{bams}{path to bam files from where the reads have to be counted}

\item{pairedEnd}{logical. whether to keep both reads of paired-end data}
}
\value{
Total counts within given ranges per BAM file.
}
\description{
Count the number of reads in a given GRanges
}
