% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_readStats}
\alias{plot_readStats}
\title{Plot read statistics from the CapSet object}
\usage{
plot_readStats(CapSet, plotType = c("numbers", "proportions"),
  outfile = NULL)
}
\arguments{
\item{CapSet}{The \code{\link{CapSet}} object}

\item{plotType}{The type of plot to make. Choose from "numbers" or "proportions". If "proportions"
is selected, the proportion of reads w.r.t total demultiplexed reads per sample
would be plotted}

\item{outfile}{Output file name. (filename extention would be used to determine type).
If outfile not specified, the plot would be retured on the screen}
}
\value{
A ggplot object, or a file. Plot showing the number/proportion of reads in each category, per sample
}
\description{
Plot read statistics from the CapSet object
}
\examples{
\dontrun{
plot_readStats(CapSet = cs, plotType = "numbers")
}
}

