% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_TSSprecision}
\alias{plot_TSSprecision}
\title{Plot precision of TSS calling between multiple samples}
\usage{
plot_TSSprecision(TSSbedFiles, sampleNames, reference, distanceCutoff = 500,
  outFile = NULL)
}
\arguments{
\item{TSSbedFiles}{BED files with TSS called by the CAGE wrapper}

\item{sampleNames}{Labels for Input samples}

\item{reference}{Reference Transcrips/Genes as a \code{\link{GRanges}} object}

\item{distanceCutoff}{Maximum distance (in base pairs) from reference TSS to plot}

\item{outFile}{Output file name (filename extention would be used to determine type).
If outfile not specified, the plot would be retured on the screen}
}
\value{
A ggplot object, or a file. Plot showing % TSS called per sample w.r.t their distance to TSS
}
\description{
Plot precision of TSS calling between multiple samples
}
\examples{
\dontrun{
library("TxDb.Dmelanogaster.UCSC.dm6.ensGene")
transcripts <- transcripts(dm6GTF)
files <- system.file("extdata", "testTSS.bed", package = "mapcapR")
plot_TSSprecision(TSSbedFiles = files, sampleNames = "testTSS",
		reference = transcripts , distanceCutoff = 500,
		outFile = "TSS_detection_precision.png")
}
}

