% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TSSprecision.R
\name{plot_TSSprecision}
\alias{plot_TSSprecision}
\title{Plot precision of TSS calling between multiple samples}
\usage{
plot_TSSprecision(TSSbedFiles, sampleNames, reference, distanceCutoff = 500,
  outFile = NULL)
}
\arguments{
\item{TSSbedFiles}{BED files with TSS called by the CAGE wrapper}

\item{sampleNames}{Labels for Input samples}

\item{reference}{Reference Transcrips/Genes GRanges}

\item{distanceCutoff}{Maximum distance upto which to plot precision}

\item{outFile}{Output png file name}
}
\value{
A plot showing % TSS called per sample w.r.t their distance to TSS
}
\description{
Plot precision of TSS calling between multiple samples
}
\examples{
\dontrun{
library("TxDb.Dmelanogaster.UCSC.dm6.ensGene")
transcripts <- transcripts(dm6GTF)
files <- system.file("extdata", c("tssclusters1.bed","tssclusters2.bed"), package = "mapcapR")
plot_TSSprecision(TSSbedFiles = files, sampleNames = c("test1", "test2"),
		reference = transcripts , distanceCutoff = 500,
		outFile = "TSS_detection_precision.png")
}
}

