% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotting_functions.R
\docType{methods}
\name{plot_TSSprecision}
\alias{plot_TSSprecision}
\alias{plot_TSSprecision,GRanges,character-method}
\alias{plot_TSSprecision,GRanges,CapSet-method}
\title{Compare the precision of TSS detection between multiple samples}
\usage{
plot_TSSprecision(reference, detectedTSS, distanceCutoff = 500,
  outFile = NULL, ...)

\S4method{plot_TSSprecision}{GRanges,character}(reference, detectedTSS,
  distanceCutoff = 500, outFile = NULL, sampleNames)

\S4method{plot_TSSprecision}{GRanges,CapSet}(reference, detectedTSS,
  distanceCutoff = 500, outFile = NULL, ...)
}
\arguments{
\item{reference}{Reference Transcrips/Genes as a \code{\link{GRanges}} object}

\item{detectedTSS}{Either a CapSet object with TSS information (after running \code{\link{detect_TSS}}
or a character vector with paths to the BED files containing detcted TSSs}

\item{distanceCutoff}{Maximum distance (in base pairs) from reference TSS to plot}

\item{outFile}{Output file name (filename extention would be used to determine type)
If outfile not specified, the plot would be returned on the screen}

\item{...}{Additional arguments}

\item{sampleNames}{Labels for input samples (in the same order as the input bed files)}
}
\value{
A ggplot object, or a file. Plot showing perent of TSS detected per sample with respect to
        their cumulative distance to TSS of the provided reference
}
\description{
Plot precision of TSS detection from multiple samples (bed files) with respect to
a given reference annotation.

Plot precision of TSS detection from multiple samples present within a
\code{\link{CapSet}} object, with respect to a given reference annotation.
}
\examples{
# load a previously saved CapSet object
cs <- exampleCSobject()
# load a txdb object
library("TxDb.Dmelanogaster.UCSC.dm6.ensGene")
seqlevelsStyle(TxDb.Dmelanogaster.UCSC.dm6.ensGene) <- "ENSEMBL"
transcripts <- transcripts(TxDb.Dmelanogaster.UCSC.dm6.ensGene)

# Plotting the precision using a pre computed set of TSS (.bed files) :

tssfile <- system.file("extdata", "testTSS.bed", package = "icetea")
plot_TSSprecision(reference = transcripts, detectedTSS = tssfile,
		sampleNames = "testTSS", distanceCutoff = 500,
		outFile = "TSS_detection_precision.png")

# Plotting the precision using a CapSet object :

library("TxDb.Dmelanogaster.UCSC.dm6.ensGene")
seqlevelsStyle(TxDb.Dmelanogaster.UCSC.dm6.ensGene) <- "ENSEMBL"
transcripts <- transcripts(TxDb.Dmelanogaster.UCSC.dm6.ensGene)

plot_TSSprecision(reference = transcripts, detectedTSS = cs,
                  outFile = "TSS_detection_precision.png")

}
