% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{sampleInfo,CapSet-method}
\alias{sampleInfo,CapSet-method}
\title{get sample information data frame}
\usage{
\S4method{sampleInfo}{CapSet}(object)
}
\arguments{
\item{object}{the \code{\link{CapSet}} object}
}
\value{
sample information data.frame
}
\description{
get sample information data frame
}
\examples{

# load a previously saved CapSet object
cs <- exampleCSobject()
# get sampleinfo
si <- sampleInfo(cs)

}
