% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{sampleInfo<-,CapSet-method}
\alias{sampleInfo<-,CapSet-method}
\title{reset sample information data frame}
\usage{
\S4method{sampleInfo}{CapSet}(object) <- value
}
\arguments{
\item{object}{The \code{\link{CapSet}} object}

\item{value}{The replacement sampleInfo data frame}
}
\description{
reset sample information data frame
}

