% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sampleInfo<-}
\alias{sampleInfo<-}
\title{reset sample information data frame}
\usage{
sampleInfo(object, ...) <- value
}
\arguments{
\item{object}{The \code{\link{CapSet}} object}

\item{...}{...}

\item{value}{new value}
}
\value{
None
}
\description{
reset sample information data frame
}
\examples{
# load a previously saved CapSet object
cs <- exampleCSobject()
# get sampleinfo
si <- sampleInfo(cs)
# modify
si$samples <- paste0("sample_", 1:6)
# replace
sampleInfo(cs) <- si

}
