% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{sampleInfo}
\alias{sampleInfo}
\title{get sample information data frame}
\usage{
sampleInfo(object, ...)
}
\arguments{
\item{object}{the \code{\link{CapSet}} object}

\item{...}{...}
}
\description{
get sample information data frame
}

