% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBAM_byRepindex.R
\name{splitBAM_byRepindex}
\alias{splitBAM_byRepindex}
\title{Split the composite BAM file using replicate indexes}
\usage{
splitBAM_byRepindex(bamFile, outfile_prefix, nthreads = 1)
}
\arguments{
\item{bamFile}{Input BAM file}

\item{outfile_prefix}{A prefix for output file (replicates IDs will be added as RR/YY)}

\item{nthreads}{Number of threads}
}
\value{
Filtered files by replicate Index
}
\description{
Split the composite BAM file using replicate indexes
}
\examples{
\dontrun{
bam <- system.file("extdata", "test_mapped.bam", package = "mapcapR")
splitBAM_byRepindex(bamFile = bam, outfile_prefix = "TEST", nthreads = 10)
}

}

